requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.87.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
passlib==1.7.4
pydantic==1.10.2
PyJWT==1.7.1
PyMySQL==1.0.2
python-decouple==3.3
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
